Fix mixin module
